# Improve variable completion performance