<!-- Anything that looks like this is a comment and can't be seen after the Pull Request is created. -->

# PR Summary
This is a small rewrite of the variable completion code to make it faster.  
While the general performance was fine, the first variable tab completion was too slow for me.  
On my PC the initial variable completion inside a long script takes about 1.8 seconds, and subsequent attempts take 50ms.  
With the new code the first attempt takes about 100ms, and drops down to around 4ms for subsequent attempts.

The logic stays mostly the same, I've simply replaced `Get-Item` calls with the equivalent internal API calls to skip the PS invocation overhead. Most of the performance improvements however come from the updated Astvisitor code. Instead of adding every variable instance to a list and later analyzing the list with a bunch of duplicates, I only look for variable assignments and save the most recent type constraint and static assignment types in a dictionary for later reference.

There are some minor functional changes that are worth discussing:
1. When analyzing script text I don't process variables after the cursor. This seems more accurate but maybe users like having all the script variables show up in the completion, even if they appear after the cursor. 

3. For commands with "OutVariable" I don't bother using the type inference to try and figure out the output type of the command and instead assume it's an arraylist. The reason for this is that PowerShell doesn't unwrap the ArrayLists created by these parameters so it's technically more accurate to list those variables as ArrayLists.  

5. Pipeline variables are only included in the completion result if the cursor is within the pipeline.
<!-- Summarize your PR between here and the checklist. -->

## PR Context
Like I said, I thought it was a bit too slow.  
Also:
Fixes https://github.com/PowerShell/PowerShell/issues/18688

<!-- Provide a little reasoning as to why this Pull Request helps and why you have opened it. -->

## PR Checklist

- [X] [PR has a meaningful title](https://github.com/PowerShell/PowerShell/blob/master/.github/CONTRIBUTING.md#pull-request---submission)
    - Use the present tense and imperative mood when describing your changes
- [X] [Summarized changes](https://github.com/PowerShell/PowerShell/blob/master/.github/CONTRIBUTING.md#pull-request---submission)
- [X] [Make sure all `.h`, `.cpp`, `.cs`, `.ps1` and `.psm1` files have the correct copyright header](https://github.com/PowerShell/PowerShell/blob/master/.github/CONTRIBUTING.md#pull-request---submission)
- [X] This PR is ready to merge and is not [Work in Progress](https://github.com/PowerShell/PowerShell/blob/master/.github/CONTRIBUTING.md#pull-request---work-in-progress).
    - If the PR is work in progress, please add the prefix `WIP:` or `[ WIP ]` to the beginning of the title (the `WIP` bot will keep its status check at `Pending` while the prefix is present) and remove the prefix when the PR is ready.
- **[Breaking changes](https://github.com/PowerShell/PowerShell/blob/master/.github/CONTRIBUTING.md#making-breaking-changes)**
    - [X] None
    - **OR**
    - [ ] [Experimental feature(s) needed](https://github.com/MicrosoftDocs/PowerShell-Docs/blob/main/reference/7.3/Microsoft.PowerShell.Core/About/about_Experimental_Features.md)
        - [ ] Experimental feature name(s): <!-- Experimental feature name(s) here -->
- **User-facing changes**
    - [X] Not Applicable
    - **OR**
    - [ ] [Documentation needed](https://github.com/PowerShell/PowerShell/blob/master/.github/CONTRIBUTING.md#pull-request---submission)
        - [ ] Issue filed: <!-- Number/link of that issue here -->
- **Testing - New and feature**
    - [ ] N/A or can only be tested interactively
    - **OR**
    - [X] [Make sure you've added a new test if existing tests do not effectively test the code changed](https://github.com/PowerShell/PowerShell/blob/master/.github/CONTRIBUTING.md#before-submitting)
- **Tooling**
    - [X] I have considered the user experience from a tooling perspective and don't believe tooling will be impacted.
    - **OR**
    - [ ] I have considered the user experience from a tooling perspective and opened an issue in the relevant tool repository. This may include:
        - [ ] Impact on [PowerShell Editor Services](https://github.com/PowerShell/PowerShellEditorServices) which is used in the [PowerShell extension](https://github.com/PowerShell/vscode-powershell) for VSCode
        (which runs in a different PS Host).
            - [ ] Issue filed: <!-- Number/link of that issue here -->
        - [ ] Impact on Completions (both in the console and in editors) - one of PowerShell's most powerful features.
            - [ ] Issue filed: <!-- Number/link of that issue here -->
        - [ ] Impact on [PSScriptAnalyzer](https://github.com/PowerShell/PSScriptAnalyzer) (which provides linting & formatting in the editor extensions).
            - [ ] Issue filed: <!-- Number/link of that issue here -->
        - [ ] Impact on [EditorSyntax](https://github.com/PowerShell/EditorSyntax) (which provides syntax highlighting with in VSCode, GitHub, and many other editors).
            - [ ] Issue filed: <!-- Number/link of that issue here -->